+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [9.0.0-alpha.74](https://github.com/Kong/kongponents/compare/v9.0.0-alpha.73...v9.0.0-alpha.74) (2023-12-11)
+
+
+### Bug Fixes
+
+* **ktruncate:** stop click propagation [KHCP-10009] ([#1904](https://github.com/Kong/kongponents/issues/1904)) ([3bec03c](https://github.com/Kong/kongponents/commit/3bec03c1bd32df49851179400dba3aa3a8386a16))
+
+# [9.0.0-alpha.73](https://github.com/Kong/kongponents/compare/v9.0.0-alpha.72...v9.0.0-alpha.73) (2023-12-08)
+
+
+### Features
+
+* **kbadge, kinputswitch, kmethodbadge:** components reskin ([#1893](https://github.com/Kong/kongponents/issues/1893)) ([07eafc8](https://github.com/Kong/kongponents/commit/07eafc8c7804c7dcc9189a85d05bcdabd8d995af)), closes [#1871](https://github.com/Kong/kongponents/issues/1871)
+
 # [9.0.0-alpha.72](https://github.com/Kong/kongponents/compare/v9.0.0-alpha.71...v9.0.0-alpha.72) (2023-12-07)